Remove AbstractAssetManager
