+++ modules/asset-manager-impl/src/test/java/org/opencastproject/assetmanager/impl/AbstractAssetManagerSelectTest.java
@@ -90,7 +90,7 @@ public class AbstractAssetManagerSelectTest extends AbstractAssetManagerTestBase
     assertFalse("Asset should not be found", am.getAsset(version, "id", "id").isSome());
     // try to find the catalog of the media package by checksum
     final MediaPackage mpCopy = MediaPackageSupport.copy(mp);
-    am.calcChecksumsForMediaPackageElements(AbstractAssetManager.assetsOnly(mpCopy));
+    am.calcChecksumsForMediaPackageElements(AssetManager.assetsOnly(mpCopy));
     assertEquals("Media package should be set up with a single catalog", 1, mpCopy.getCatalogs().length);
     final String checksum = mpCopy.getCatalogs()[0].getChecksum().toString();
     assertTrue("Media package element should be retrievable by checksum", am.getDb().findAssetByChecksum(checksum).isSome());
